refactor: moved advert related hooks to its own entity
