+++ packages/api/src/hooks/p2p/__tests__/useAdvertiserAdverts.spec.tsx
@@ -1,5 +1,5 @@
 import { renderHook } from '@testing-library/react-hooks';
-import useAdvertiserAdverts from '../useAdvertiserAdverts';
+import useAdvertiserAdverts from '../entity/advert/p2p-advertiser-adverts/useAdvertiserAdverts';
 import useInfiniteQuery from '../../../useInfiniteQuery';
 import APIProvider from '../../../APIProvider';
 import React from 'react';
