+++ packages/api/src/hooks/p2p/entity/advert/index.ts
@@ -1,0 +1,2 @@
+export * as advert from './p2p-advert';
+export * as advertiserAdverts from './p2p-advertiser-adverts';
