+++ packages/api/src/hooks/p2p/entity/advert/p2p-advert/index.ts
@@ -1,0 +1,5 @@
+export { default as useGet } from './useAdvertInfo';
+export { default as useGetList } from './useAdvertList';
+export { default as useCreate } from './useAdvertCreate';
+export { default as useUpdate } from './useAdvertUpdate';
+export { default as useDelete } from './useAdvertDelete';
