+++ packages/api/src/hooks/p2p/entity/advert/p2p-advertiser-adverts/index.ts
@@ -1,0 +1,1 @@
+export { default as useGet } from './useAdvertiserAdverts';
