+++ packages/api/src/hooks/p2p/entity/advert/p2p-advertiser-adverts/useAdvertiserAdverts.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
-import useInfiniteQuery from '../../useInfiniteQuery';
-import useAuthorize from '../useAuthorize';
+import useInfiniteQuery from '../../../../../useInfiniteQuery';
+import useAuthorize from '../../../../useAuthorize';
 
 /** This custom hook returns a list of adverts under the current active client. */
 const useAdvertiserAdverts = (
